Cleanup obsolete code (#555)
