+++ optimum/neuron/accelerate/state.py
@@ -32,7 +32,7 @@ from accelerate.utils import (
     parse_choice_from_env,
     parse_flag_from_env,
 )
-from accelerate.utils.dataclasses import FullyShardedDataParallelPlugin, SageMakerDistributedType
+from accelerate.utils.dataclasses import SageMakerDistributedType
 
 from ...utils import logging
 from ..utils import is_neuronx_distributed_available, is_torch_xla_available
@@ -41,7 +41,7 @@ from ..utils.torch_xla_and_neuronx_initialization import (
     set_common_flags,
     set_neuron_cc_flags_for_torch_amp,
 )
-from .utils import NeuronDistributedType, NeuronFullyShardedDataParallelPlugin
+from .utils import NeuronDistributedType
 from .utils.dataclasses import AutocastBackend, ModelParallelismPlugin
 
 
@@ -201,7 +201,7 @@ class NeuronPartialState(PartialState):
         self.fork_launched = parse_flag_from_env("FORK_LAUNCHED", 0)
 
     def wait_for_everyone(self):
-        if self.distributed_type in [NeuronDistributedType.XLA_FSDP, NeuronDistributedType.MODEL_PARALLELISM]:
+        if self.distributed_type is NeuronDistributedType.MODEL_PARALLELISM:
             xm.rendezvous("accelerate.utils.wait_for_everyone")
         else:
             super().wait_for_everyone()
@@ -303,17 +303,7 @@ class NeuronAcceleratorState(AcceleratorState):
                         pipeline_model_parallel_size=self.mp_plugin.pipeline_parallel_size,
                     )
 
-                if os.environ.get("ACCELERATE_USE_FSDP", "false") == "true":
-                    self.distributed_type = NeuronDistributedType.XLA_FSDP
-                    if self._mixed_precision != "no":
-                        # TODO: do we need that?
-                        fsdp_plugin.set_mixed_precision(self._mixed_precision)
-                    if isinstance(fsdp_plugin, FullyShardedDataParallelPlugin) and not isinstance(
-                        fsdp_plugin, NeuronFullyShardedDataParallelPlugin
-                    ):
-                        fsdp_plugin.__class__ = NeuronFullyShardedDataParallelPlugin
-                    self.fsdp_plugin = fsdp_plugin
-            elif os.environ.get("ACCELERATE_USE_DEEPSPEED", "false") == "true" and not cpu:
+            if os.environ.get("ACCELERATE_USE_DEEPSPEED", "false") == "true" and not cpu:
                 self.deepspeed_plugin = deepspeed_plugin
             elif self.distributed_type == DistributedType.MULTI_GPU:
                 if os.environ.get("ACCELERATE_USE_FSDP", "false") == "true":
