+++ optimum/neuron/accelerate/utils/__init__.py
@@ -17,6 +17,5 @@ from .dataclasses import (
     AutocastBackend,
     ModelParallelismPlugin,
     NeuronDistributedType,
-    NeuronFullyShardedDataParallelPlugin,
 )
 from .misc import get_tied_parameters_dict, patch_accelerate_is_tpu_available, tie_parameters
